[IOTDB-279] Merge TsDigest into Statistics
