+++ server/src/main/java/org/apache/iotdb/db/query/reader/resourceRelated/UnseqResourceMergeReader.java
@@ -35,7 +35,7 @@ import org.apache.iotdb.db.query.reader.chunkRelated.ChunkReaderWrap;
 import org.apache.iotdb.db.query.reader.universal.PriorityMergeReader;
 import org.apache.iotdb.db.utils.QueryUtils;
 import org.apache.iotdb.tsfile.file.metadata.ChunkMetaData;
-import org.apache.iotdb.tsfile.file.metadata.TsDigest.StatisticType;
+import org.apache.iotdb.tsfile.file.metadata.statistics.Statistics;
 import org.apache.iotdb.tsfile.read.TsFileSequenceReader;
 import org.apache.iotdb.tsfile.read.common.Path;
 import org.apache.iotdb.tsfile.read.controller.ChunkLoaderImpl;
@@ -102,8 +102,8 @@ public class UnseqResourceMergeReader extends PriorityMergeReader {
           ByteBuffer maxValue = null;
           ByteBuffer[] statistics = chunkMetaData.getDigest().getStatistics();
           if (statistics != null) {
-            minValue = statistics[StatisticType.min_value.ordinal()]; // note still CAN be null
-            maxValue = statistics[StatisticType.max_value.ordinal()]; // note still CAN be null
+            minValue = statistics[Statistics.StatisticType.min_value.ordinal()]; // note still CAN be null
+            maxValue = statistics[Statistics.StatisticType.max_value.ordinal()]; // note still CAN be null
           }
 
           DigestForFilter digest = new DigestForFilter(chunkMetaData.getStartTime(),
