+++ tsfile/src/main/java/org/apache/iotdb/tsfile/read/reader/series/FileSeriesReaderWithFilter.java
@@ -22,7 +22,7 @@ import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.List;
 import org.apache.iotdb.tsfile.file.metadata.ChunkMetaData;
-import org.apache.iotdb.tsfile.file.metadata.TsDigest.StatisticType;
+import org.apache.iotdb.tsfile.file.metadata.statistics.Statistics;
 import org.apache.iotdb.tsfile.read.common.Chunk;
 import org.apache.iotdb.tsfile.read.controller.IChunkLoader;
 import org.apache.iotdb.tsfile.read.filter.DigestForFilter;
@@ -55,8 +55,8 @@ public class FileSeriesReaderWithFilter extends FileSeriesReader {
     ByteBuffer maxValue = null;
     ByteBuffer[] statistics = chunkMetaData.getDigest().getStatistics();
     if (statistics != null) {
-      minValue = statistics[StatisticType.min_value.ordinal()]; // note still CAN be null
-      maxValue = statistics[StatisticType.max_value.ordinal()]; // note still CAN be null
+      minValue = statistics[Statistics.StatisticType.min_value.ordinal()]; // note still CAN be null
+      maxValue = statistics[Statistics.StatisticType.max_value.ordinal()]; // note still CAN be null
     }
 
     DigestForFilter digest = new DigestForFilter(chunkMetaData.getStartTime(),
